Move XorName to a separate file so its easier to find
